Video Player Updates
Reorganized the DynamicDataSetLoader so that it's only updating when the helper.txt file in the cloud is new. When updating, it pulls in the xml, parses it for the target names, finds the associated txt files in the cloud, and then downloads the .dat file.

Also updated the metadataParse.cs to make sure that the video doesn't play until its prepared. This causes the load time to increase, but ensures a smooth playback.

To account for the long load time, I started to add the preview video that would switch to the full video when it's prepared, but that functionality still needs additional work.